添加接口
使有关材质的类实现新增的ITexture接口，增强封装性，性能暂时无明显下降
--------------------性能--------------------
驱动程序版本: 347.09
帧数(f/s): 190

+++ GaLiong/BMP.h
@@ -7,8 +7,8 @@ _L_BEGIN
 class _L_ BMP
 {
 public:
-	static bool InitHeader(ifstream &stream, Size &size, int &length);
-	static unsigned char *ReadData(ifstream &stream, int length);
+	static bool InitHeader(ifstream &stream, Size &size, unsigned long &length);
+	static unsigned char *ReadData(ifstream &stream, unsigned long length);
 	static void Automatic(ifstream &stream, wchar_t *path, Texture &texture);
 	static void Automatic(ifstream &stream, wchar_t *path, TextureBuffer &texture);
 	static void Automatic_Unsafe(ifstream &stream, wchar_t *path, Texture &texture);
